+++ glide-core/redis-rs/redis/src/lib.rs
@@ -339,7 +339,9 @@ assert_eq!(result, Ok(("foo".to_string(), b"bar".to_vec())));
 pub use crate::client::Client;
 pub use crate::client::GlideConnectionOptions;
 pub use crate::client::IAMTokenProvider;
-pub use crate::cmd::{cmd, fenced_cmd, pack_command, pipe, Arg, Cmd, Iter};
+pub use crate::cmd::{
+    cmd, fenced_cmd, pack_command, pipe, Arg, Cmd, Iter, PHASE_QUEUED, PHASE_SENT,
+};
 pub use crate::commands::{
     Commands, ControlFlow, Direction, LposOptions, PubSubCommands, SetOptions,
 };